fix:1.Add Blueprint. 2.Add animation effects. 3.Conceal confidential messages.
