+++ .gitignore
@@ -1,0 +1,4 @@
+# ignore .env file
+.env
+
+.__pycache__
