+++ static/JS/order.js
@@ -1,5 +1,4 @@
 //取得付款資訊
-
 const btnBackIndex = document.querySelector(".btnBackIndex");
 btnBackIndex.addEventListener("click", () => {
   window.location.href = `/`;
@@ -9,31 +8,31 @@ const orderContact = document.querySelector(".thank_contact");
 const urlParams = new URLSearchParams(window.location.search);
 const orderNumber = urlParams.get("number");
 console.log(orderNumber);
-//建立新的付款訂單
-const getOrder = () => {
-  fetch(`/api/order/${orderNumber}`, {
-    method: "GET",
-    headers: {
-      Authorization: "Bearer " + token,
-      "Content-Type": "application/json",
-    },
-  })
-    .then((respose) => respose.json())
-    .then((result) => {
-      const order = result.data;
-      const trip = order.trip;
-      const attraction = order.trip.attraction;
-      console.log(order);
-      const number = order.number;
-      const name = attraction.name;
-      const date = trip.date;
-      const time = trip.time;
-      const price = order.price;
-      const contactName = order.contact.name;
-      const contactEmail = order.contact.email;
-      const contactPhone = order.contact.phone;
 
-      const renderInfo = `
+async function getOrder() {
+  try {
+    const response = await fetch(`/api/order/${orderNumber}`, {
+      method: "GET",
+      headers: {
+        Authorization: "Bearer " + token,
+        "Content-Type": "application/json",
+      },
+    });
+    const result = await response.json();
+    const order = result.data;
+    const trip = order.trip;
+    const attraction = order.trip.attraction;
+    console.log(order);
+    const number = order.number;
+    const name = attraction.name;
+    const date = trip.date;
+    const time = trip.time;
+    const price = order.price;
+    const contactName = order.contact.name;
+    const contactEmail = order.contact.email;
+    const contactPhone = order.contact.phone;
+
+    const renderInfo = `
       <div class="thank_title2" id="thank_number">訂單編號：${number}</div>
       <div class="thank_title2">
         景點：
@@ -53,7 +52,7 @@ const getOrder = () => {
         費用：
         <div class="thank_text" id="thank_price">新台幣${price}元</div>
       </div>`;
-      const renderContact = `
+    const renderContact = `
       <div class="thank_title2">
       聯絡姓名：
       <div class="thank_text" id="thank_contact_name">${contactName}</div>
@@ -66,12 +65,11 @@ const getOrder = () => {
       Email：
       <div class="thank_text" id="thank_contact_phone">${contactEmail}</div>
     </div>`;
-      orderInfo.insertAdjacentHTML("beforeend", renderInfo);
-      orderContact.insertAdjacentHTML("beforeend", renderContact);
-    })
-    .catch((error) => {
-      // console.log(error);
-      console.log("API呼叫失敗:" + error.message);
-    });
-};
+    orderInfo.insertAdjacentHTML("beforeend", renderInfo);
+    orderContact.insertAdjacentHTML("beforeend", renderContact);
+  } catch (error) {
+    // console.log(error);
+    console.log("API呼叫失敗:" + error.message);
+  }
+}
 getOrder();
