+++ static/css/index.css
@@ -8,7 +8,7 @@ body {
   display: grid;
   font-family: "Noto Sans TC";
   font-style: normal;
-  grid-template-rows: 54px 320px auto 104px;
+  grid-template-rows: 54px 320px auto minmax(104px, auto);
   justify-self: center;
 }
 
@@ -18,7 +18,7 @@ nav {
   grid-template-columns: auto minmax(300px, 1200px) auto;
   position: sticky;
   background-color: #ffffff;
-  z-index: 10;
+  z-index: 200;
   top: 0;
 }
 
@@ -48,20 +48,24 @@ nav > .nav_box > .nav_items > .nav_item {
   line-height: 13px;
   color: #666666;
   border: 0;
-  /*去邊*/
   background-color: #ffffff;
   padding-right: 10px;
+  cursor: pointer;
+}
+
+nav > .nav_box > .nav_items > .nav_item:hover {
+  color: #448899;
 }
 
 header {
   display: grid;
   grid-row: 2/3;
-  background-image: url(../css/picture/welcome.png);
+  background-image: url(../css/picture/home_cover.png);
   grid-template-columns: auto minmax(300px, 1200px) auto;
-  /* justify-content: center; */
   align-content: center;
   background-repeat: no-repeat;
   background-size: cover;
+  background-position: center;
 }
 
 header > .welcome {
@@ -110,6 +114,7 @@ header > .welcome > .welcome_search > .search_btn {
   background: #448899;
   border-radius: 0px 5px 5px 0px;
   position: absolute;
+  cursor: pointer;
 }
 
 header > .welcome > .welcome_search > .category_list {
@@ -153,6 +158,10 @@ main > .index_card {
   display: grid;
   grid-template-rows: 157px 40px 43px;
   border: 1px solid #e8e8e8;
+  border-radius: 5px 5px 5px 5px;
+  box-shadow: 1px 1px 20px 2px rgb(0 0 0 / 10%);
+  overflow: hidden;
+  position: relative;
 }
 
 main > .index_card > .index_img {
@@ -160,6 +169,18 @@ main > .index_card > .index_img {
   grid-area: 1 / 1 / 3 / 2;
   background-size: cover;
   background-position: center;
+  transform: scale(1, 1);
+  opacity: 90%;
+  transition: all 1s ease-out;
+  cursor: pointer;
+  object-fit: cover;
+  border-radius: 5px 5px 0px 0px;
+}
+
+main > .index_card > .index_img:hover {
+  transform: scale(1.2, 1.2);
+  opacity: 100%;
+  z-index: 1;
 }
 
 main > .index_card > .index_name {
@@ -174,6 +195,7 @@ main > .index_card > .index_name {
   color: #ffffff;
   align-items: center;
   padding-left: 15px;
+  z-index: 2;
 }
 
 main > .index_card > .index_inf {
@@ -181,6 +203,8 @@ main > .index_card > .index_inf {
   grid-row: 3/4;
   background: #ffffff;
   grid-template-columns: 1fr 1fr;
+  border-radius: 0px 0px 5px 5px;
+  z-index: 2;
 }
 
 main > .index_card > .index_inf > .index_mrt {
@@ -214,7 +238,7 @@ footer {
   line-height: 13px;
   justify-content: center;
   align-content: center;
-  height: 104px;
+  /* height: 104px; */
 }
 
 @media (max-width: 1200px) {
