+++ templates/attraction.html
@@ -152,6 +152,14 @@
         </div>
       </form>
     </div>
+    <div class="notify_box">
+      <div class="notify_top"></div>
+      <button class="btn_close_notify">
+        <img src="../static/css/picture/icon_close.png" />
+      </button>
+      <div class="notify_title">系統通知</div>
+      <div class="notify_message"></div>
+    </div>
     <div class="overlay"></div>
     <script src="../static/JS/attraction.js"></script>
     <script src="../static/JS/user.js"></script>
